reorganization and clearer types for sessions
