+++ media_recommendation_engine/src/routes/library.rs
@@ -19,8 +19,9 @@ use crate::{
     state::{AppResult, AppState, Shutdown},
     utils::{
         frontend_redirect, frontend_redirect_explicit,
+        streaming::StreamingSessions,
         templates::{GridElement, LargeImage, Library, PreviewTemplate},
-        HXTarget, StreamingSessions,
+        HXTarget,
     },
 };
