+++ media_recommendation_engine/src/utils/streaming/mod.rs
@@ -1,0 +1,3 @@
+mod communication;
+mod session;
+pub use session::{Session, StreamingSessions};
